Route class moved into /routes/models/Route.tsx file
